Added event management in search player.
The search player component emits an event when player is selected in suggest list, or emit an 'empty' player event when field is cleared.
The search team component listens to the event, and stores players (to lookup for the temp object after all).
Added a team service component to get the list all players involded in the match and remove them from the suggestion list.